shpfloor: use catch_error decorator in zp, spt
