+++ shopfloor/tests/test_single_pack_transfer_base.py
@@ -1,6 +1,8 @@
 # Copyright 2020 Camptocamp SA (http://www.camptocamp.com)
 # Copyright 2020 Akretion (http://www.akretion.com)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+from odoo.tests.common import Form
+
 from .common import CommonCase
 
 
@@ -28,3 +30,19 @@ class SinglePackTransferCommonBase(CommonCase):
         self.service = self.get_service(
             "single_pack_transfer", menu=self.menu, profile=self.profile
         )
+
+    @classmethod
+    def _create_initial_move(cls, lines):
+        """Create the move to satisfy the pre-condition before /start"""
+        picking_form = Form(cls.env["stock.picking"])
+        picking_form.picking_type_id = cls.picking_type
+        picking_form.location_id = cls.stock_location
+        picking_form.location_dest_id = cls.shelf2
+        for line in lines:
+            with picking_form.move_ids_without_package.new() as move:
+                move.product_id = line[0]
+                move.product_uom_qty = line[1]
+        picking = picking_form.save()
+        picking.action_confirm()
+        picking.action_assign()
+        return picking
